Add initial implementation of AsyncLoader
The idea is that AsyncLoader allows for a cleaner usage of WeekView when using a reactive architecture. The sample app's AsyncActivity shows an example using a ViewModel that fetches events from an API. Instead of having to provide an OnMonthChangeListener and using WeekView.notifyDataSetChanged(), it simply calls WeekView.submit() when the events are fetched.

If the ViewModel should be notified whenever WeekView requires more events (e.g. due to the user scrolling), WeekView offers a onLoadMore() callback to do so. It operates like an OnMonthChangeListener, but doesn't return anything.