Deprecate WeekViewDisplayable and move WeekViewEntity creation to adapter
This commit deprecates WeekViewDisplayable. Instead of requiring data classes to implement this interface for constructing a WeekViewEntity (or the previous WeekViewEvent), this construction now occurs in the adapter’s onCreateEntity() method.

This enables a better separation of concerns and makes it possible to pass a Context to the construction site, allowing you to resolve resources directly. This can be helpful when deciding on which color to use depending on the device’s current theme.